requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
playwright==1.46.0
yt-dlp==2025.1.15
tenacity==9.0.0

google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

pytest==8.3.2
